fix: float to double

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/loanschedule/domain/remoteschedulegenerator/Charge.java
@@ -5,8 +5,8 @@ import java.util.Date;
 public class Charge {
   private String id;
   private Date createdTimestamp;
-  private float amount;
-  private float percentage;
+  private double amount;
+  private double percentage;
   private ChargePercentageBase percentageBase;
   private Frequency chargeFrequency;
   private ChargeAppliesOn appliesOn;
