+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/loanschedule/domain/remoteschedulegenerator/InstallmentComponent.java
@@ -1,44 +1,44 @@
 package org.apache.fineract.portfolio.loanaccount.loanschedule.domain.remoteschedulegenerator;
 
 public class InstallmentComponent {
-  private float principal;
-  private float interest;
-  private float fee;
-  private float penalty;
+  private double principal;
+  private double interest;
+  private double fee;
+  private double penalty;
 
   public InstallmentComponent() {
     super();
   }
 
-  public float getPrincipal() {
+  public double getPrincipal() {
     return principal;
   }
 
-  public float getPenalty() {
+  public double getPenalty() {
     return penalty;
   }
 
-  public void setPenalty(float penalty) {
+  public void setPenalty(double penalty) {
     this.penalty = penalty;
   }
 
-  public float getFee() {
+  public double getFee() {
     return fee;
   }
 
-  public void setFee(float fee) {
+  public void setFee(double fee) {
     this.fee = fee;
   }
 
-  public float getInterest() {
+  public double getInterest() {
     return interest;
   }
 
-  public void setInterest(float interest) {
+  public void setInterest(double interest) {
     this.interest = interest;
   }
 
-  public void setPrincipal(float principal) {
+  public void setPrincipal(double principal) {
     this.principal = principal;
   }
 }
